Test rename paths
