+++ MANIFEST.in
