--- simulations/agents/__init__.py
@@ -1,10 +1,0 @@
-"""
-Simulation agent definitions
-"""
-
-from simulations.agents.borrow_agent import BorrowAgent
-from simulations.agents.liquidation_agent import (
-    AdversarialLiquidationAgent,
-    LiquidationAgent,
-)
-from simulations.agents.uniswap_agent import DummyUniswapAgent, UniswapAgent
