+++ verbs_examples/aave/sim.py
@@ -46,15 +46,18 @@ from typing import List
 
 import verbs
 
-from simulations import abis
-from simulations.agents import (
+from verbs_examples import abis
+from verbs_examples.agents import (
     AdversarialLiquidationAgent,
     BorrowAgent,
     DummyUniswapAgent,
     LiquidationAgent,
     UniswapAgent,
 )
-from simulations.utils.erc20 import mint_and_approve_dai, mint_and_approve_weth
+from verbs_examples.utils.erc20 import (
+    mint_and_approve_dai,
+    mint_and_approve_weth,
+)
 
 PATH = Path(__file__).parent
