+++ verbs_examples/abis/MockAggregator.abi
