+++ verbs_examples/abis/UniswapAggregator.sol
