+++ verbs_examples/utils/sol_compiler.py
